refactor(search): move SqlQueryBuilder to :feature:search module
